+++ description.R
@@ -1,5 +1,5 @@
 pkgs <- c(
-    "tidyverse", "patchwork", "cowplot", "viridis", "ggrepel", "wesanderson", "RColorBrewer")
+    "tidyverse", "patchwork", "cowplot", "viridis", "ggrepel", "wesanderson", "RColorBrewer", "ggplotify")
 
 for(x in pkgs){
     usethis::use_package(x, type = "depends")} #, type = "depends"